migrate to SDL3

+++ include/graphics/Circle.hpp
@@ -2,7 +2,7 @@
 #define __MB_CIRCLE_H__
 #include <tuple>
 #include <vector>
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 #include <system/Math.hpp>
 #include <graphics/Renderable.hpp>
 
@@ -15,14 +15,14 @@ class Circle : public Renderable
 private:
     friend Renderer;
 
-    int mRadius { 0 };
+    float mRadius { 0 };
     bool mDrawFilled { true };
     int mThickness { 1 };
 public:
     void SetThickness(int t) { mThickness = t; } 
     void SetFilled(bool f) { mDrawFilled = f; }
     void Draw(SDL_Renderer*, Camera*) override;
-    void SetRadius(int r) { mRadius = r; }
+    void SetRadius(float r) { mRadius = r; }
 
     Circle(int r) { mRadius = r; }
